Refactor up
