+++ frontend/buildings/reducers/buildingsListReducer.js
@@ -1,42 +1,12 @@
+import { filterBuildings } from "../selectors";
+
 const DEFAULT_STATE = {
   fetching: true,
   buildings: [],
   filteredBuildings: []
 };
 
-const compactFilter = element => {
-  return !!element;
-};
-
-const toString = value => {
-  return value.toString().toLowerCase();
-};
-
-const filterBuildings = (buildings, filter) => {
-  if (filter == "") {
-    return buildings;
-  }
-
-  const filtered = buildings.filter(building => {
-    const values = [
-      building.number,
-      building.address,
-      building.neighborhood,
-      building.name
-    ];
-
-    const string = values
-      .filter(compactFilter)
-      .map(toString)
-      .join(" ");
-
-    return string.search(filter) >= 0;
-  });
-
-  return filtered;
-};
-
-export default function buildingsListReducer(state = {}, action) {
+export default function buildingsListReducer(state = DEFAULT_STATE, action) {
   if (action.type == "FETCHING_BUILDINGS") {
     return {
       ...state,
@@ -48,6 +18,7 @@ export default function buildingsListReducer(state = {}, action) {
     return {
       ...state,
       buildings: action.buildings,
+      filteredBuildings: action.buildings,
       fetching: false
     };
   }
